[Forge] Separate network latency, bandwidth and partition tests
