+++ testsuite/testcases/src/lib.rs
@@ -4,7 +4,9 @@
 pub mod compatibility_test;
 pub mod fixed_tps_test;
 pub mod gas_price_test;
-pub mod network_chaos_test;
+pub mod network_bandwidth_test;
+pub mod network_latency_test;
+pub mod network_partition_test;
 pub mod partial_nodes_down_test;
 pub mod performance_test;
 pub mod reconfiguration_test;
